Removed EmployeeId from models, keeping only approved leaverequests and sickrequests

+++ EmployeeManagement/Repositories/EmployeeRepository.cs
@@ -1,12 +1,14 @@
 ﻿using EmployeeManagement.Data;
 using EmployeeManagement.Models;
 using Microsoft.EntityFrameworkCore;
+using static EmployeeManagement.Enums.EnumTypes;
 
 namespace EmployeeManagement.Repositories
 {
     public class EmployeeRepository : IEmployeeRepository
     {
         private readonly AppDbContext _context;
+
         public EmployeeRepository(AppDbContext context)
         {
             _context = context;
@@ -40,12 +42,47 @@ namespace EmployeeManagement.Repositories
             await _context.SaveChangesAsync();
         }
 
-        public async Task DeleteEmployeeByIdAsync(Employee employee)
-        {
-            _context.Employees.Remove(employee);
-            await _context.SaveChangesAsync();
-        }
+        public async Task DeleteEmployeeByIdAsync(int employeeId)
+        {  
+            var employee = await _context.Employees
+            .Include(e => e.LeaveRequests)
+            .Include(e => e.SickLeaves)
+            .FirstOrDefaultAsync(e => e.EmployeeId == employeeId);
 
+            if (employee != null)
+            {
+                // Handle leave requests
+                foreach (var request in employee.LeaveRequests.ToList())
+                {
+                    if (request.LeaveStatus == LeaveStatus.Approved)
+                    {
+                        request.Employee = null;
+                    }
+                    else
+                    {
+                        // Remove non-approved requests
+                        _context.LeaveRequests.Remove(request);
+                    }
+                }
+
+                foreach (var request in employee.SickLeaves.ToList())
+                {
+                    if (request.LeavStatus == LeaveStatus.Approved)
+                    {
+                        request.Employee = null;
+                    }
+                    else
+                    {
+                        // Remove non-approved requests
+                        _context.SickLeaves.Remove(request);
+                    }
+                }
+
+                _context.Employees.Remove(employee);
+                await _context.SaveChangesAsync();
+            }
+            
+        }
         public async Task UpdateEmployeeAsync(Employee employee)
         {
             _context.Employees.Update(employee);
